% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_irt.R
\name{compute_recovery_metrics}
\alias{compute_recovery_metrics}
\title{Compute parameter recovery metrics}
\usage{
compute_recovery_metrics(post_means, truth)
}
\arguments{
\item{post_means}{List. Posterior means generated by \code{\link{posterior_means_from_fit}}.}

\item{truth}{List. True parameter values generated by \code{\link{simulate_sgp_irt_binary}}.}
}
\value{
A list containing the RMSE for theta, alpha, beta, and gamma.
}
\description{
Calculates the Root Mean Square Error (RMSE) between the true simulated parameters
and the estimated posterior means.
}
