% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_irt.R
\name{posterior_mean_probabilities}
\alias{posterior_mean_probabilities}
\title{Compute posterior mean predictive probabilities}
\usage{
posterior_mean_probabilities(fit_obj, truth, ndraws = 250)
}
\arguments{
\item{fit_obj}{List. The output object from \code{\link{fit_sgp_irt_binary}}.}

\item{truth}{List. True dataset object generated by \code{\link{simulate_sgp_irt_binary}}.}

\item{ndraws}{Integer. Number of posterior draws to use for the calculation. Default is 250.}
}
\value{
A numeric vector of length N containing the posterior mean probability of y = 1.
}
\description{
Calculates the expected probability of a positive response for each observation
by integrating over the posterior distribution of the model parameters.
}
