% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_irt.R
\name{posterior_means_from_fit}
\alias{posterior_means_from_fit}
\title{Extract posterior means from the fitted model}
\usage{
posterior_means_from_fit(fit_obj)
}
\arguments{
\item{fit_obj}{List. The output object from \code{\link{fit_sgp_irt_binary}}.}
}
\value{
A list of numeric vectors containing the posterior means for each parameter.
}
\description{
Calculates the column means for the primary parameters (theta, alpha, beta, gamma)
from the posterior draws.
}
